Az bot: additional refactoring and create BotJsonFormatter, WebAppOperations, BotTemplateDeployer
