+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/__init__.py
@@ -21,7 +21,7 @@ class BotServiceCommandsLoader(AzCommandsLoader):
                                                        suppress_extension=ModExtensionSuppress(
                                                            __name__,
                                                            'botservice',
-                                                           '0.4.0',
+                                                           '0.4.1',
                                                            reason='These commands are now in the CLI',
                                                            recommend_remove=True))
