+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/bot_template_deployer.py
@@ -1,0 +1,220 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+import json
+import os
+import re
+import requests
+import string
+import random
+
+from knack.util import CLIError
+from azure.cli.core.profiles import ResourceType, get_sdk
+from azure.cli.core.commands.client_factory import get_mgmt_service_client
+from azure.cli.core.util import get_file_json, shell_safe_json_parse
+from .bot_json_formatter import BotJsonFormatter
+
+
+class BotTemplateDeployer:
+    # Function App
+    function_template_name = 'functionapp.template.json'
+
+    # TODO: Move into Azure-specific folder
+    @staticmethod
+    def _get_app_insights_location(key):
+        region_map = {
+            'australiaeast': 'southeastasia',
+            'australiacentral': 'southeastasia',
+            'australiacentral2': 'southeastasia',
+            'australiasoutheast': 'southeastasia',
+            'eastasia': 'southeastasia',
+            'southeastasia': 'westus',
+            'eastus': 'eastus',
+            'eastus2': 'eastus',
+            'southcentralus': 'southcentralus',
+            'westcentralus': 'westus2',
+            'westus': 'westus2',
+            'westus2': 'westus2',
+            'brazilsouth': 'southcentralus',
+            'centralus': 'southcentralus',
+            'northcentralus': 'southcentralus',
+            'japanwest': 'southeastasia',
+            'japaneast': 'southeastasia',
+            'southindia': 'southeastasia',
+            'centralindia': 'southeastasia',
+            'westindia': 'southeastasia',
+            'canadacentral': 'southcentralus',
+            'canadaeast': 'eastus',
+            'koreacentral': 'southeastasia',
+            'koreasouth': 'southeastasia',
+            'northeurope': 'northeurope',
+            'westeurope': 'westeurope',
+            'uksouth': 'westeurope',
+            'ukwest': 'westeurope',
+            'francecentral': 'westeurope',
+            'francesouth': 'westeurope'
+        }
+        return region_map[key]
+
+    # TODO: Tear this out into separate class that deploys bots
+    @staticmethod
+    def deploy_arm_template(cli_ctx, resource_group_name,  # pylint: disable=too-many-arguments
+                            template_file=None, deployment_name=None,
+                            parameters=None, mode=None):
+        DeploymentProperties, _ = get_sdk(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES,
+                                          'DeploymentProperties', 'TemplateLink', mod='models')
+        template = None
+        template = get_file_json(template_file, preserve_order=True)
+        template_obj = template
+
+        template_obj['resources'] = template_obj.get('resources', [])
+        parameters = BotTemplateDeployer._process_parameters(parameters) or {}
+
+        import json
+        template = json.loads(json.dumps(template))
+        parameters = json.loads(json.dumps(parameters))
+
+        properties = DeploymentProperties(template=template, template_link=None,
+                                          parameters=parameters, mode=mode)
+
+        smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
+        return smc.deployments.create_or_update(resource_group_name, deployment_name, properties, raw=False)
+
+    @staticmethod
+    def create_app(cmd, client, resource_group_name, resource_name, description, kind, appid, password,
+                   storageAccountName,
+                   location, sku_name, appInsightsLocation, language, version):
+
+        # Based on sdk version, language and kind, select the appropriate zip url containing starter bot source
+        if version == 'v3':
+            if kind == 'function':
+                template_name = 'functionapp.template.json'
+                if language == 'Csharp':
+                    zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/csharp-abs-functions_emptybot.zip'
+                else:
+                    zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/node.js-abs-functions_emptybot_funcpack.zip'  # pylint: disable=line-too-long
+
+            else:
+                kind = 'sdk'
+                template_name = 'webapp.template.json'
+                if language == 'Csharp':
+                    zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/csharp-abs-webapp_simpleechobot_precompiled.zip'  # pylint: disable=line-too-long
+                else:
+                    zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/node.js-abs-webapp_hello-chatconnector.zip'  # pylint: disable=line-too-long
+        elif version == 'v4':
+            if kind == 'function':
+                raise CLIError('Function bot creation is not supported for v4 bot sdk.')
+
+            else:
+                kind = 'sdk'
+                template_name = 'webappv4.template.json'
+                if language == 'Csharp':
+                    zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/csharp-abs-webapp-v4_echobot_precompiled.zip'  # pylint: disable=line-too-long
+                else:
+                    zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/node.js-abs-webapp-v4_echobot.zip'  # pylint: disable=line-too-long
+
+        # Storage prep
+        create_new_storage = False
+        if not storageAccountName:
+            create_new_storage = True
+            storageAccountName = re.sub(r'[^a-z0-9]', '', resource_name[:10] +
+                                        ''.join(
+                                            random.choice(string.ascii_lowercase + string.digits) for _ in range(4)))
+            site_name = re.sub(r'[^a-z0-9]', '', resource_name[:15] +
+                               ''.join(random.choice(string.ascii_lowercase + string.digits) for _ in range(4)))
+
+        # Application insights prep
+        appInsightsLocation = BotTemplateDeployer._get_app_insights_location(location.lower().replace(' ', ''))
+
+        # ARM Template parameters
+        paramsdict = {
+            "location": location,
+            "kind": kind,
+            "sku": sku_name,
+            "siteName": site_name,
+            "appId": appid,
+            "appSecret": password,
+            "storageAccountResourceId": "",
+            "serverFarmId": "/subscriptions/{0}/resourceGroups/{1}/providers/Microsoft.Web/serverfarms/{2}".format(
+                client.config.subscription_id, resource_group_name, resource_name),
+            "zipUrl": zip_url,
+            "createNewStorage": create_new_storage,
+            "storageAccountName": storageAccountName,
+            "botEnv": "prod",
+            "useAppInsights": True,
+            "appInsightsLocation": appInsightsLocation,
+            "createServerFarm": True,
+            "serverFarmLocation": location.lower().replace(' ', ''),
+            "azureWebJobsBotFrameworkDirectLineSecret": "",
+            "botId": resource_name
+        }
+        if description:
+            paramsdict['description'] = description
+        if template_name == 'webappv4.template.json':
+            bot_encryption_key = BotTemplateDeployer.get_bot_file_encryption_key()
+            paramsdict['botFileEncryptionKey'] = bot_encryption_key
+        params = {k: {'value': v} for k, v in paramsdict.items()}
+
+        # TODO concate 'templates' to file path
+        dir_path = os.path.dirname(os.path.realpath(__file__))
+        dir_path = dir_path + '/templates/'
+        # Deploy ARM template
+        deploy_result = BotTemplateDeployer.deploy_arm_template(
+            cli_ctx=cmd.cli_ctx,
+            resource_group_name=resource_group_name,
+            template_file=os.path.join(dir_path, template_name), # This path is incorrect.
+            parameters=[[json.dumps(params)]],
+            deployment_name=resource_name,
+            mode='Incremental'
+        )
+
+        deploy_result.wait()
+        return BotJsonFormatter.create_bot_json(cmd, client, resource_group_name, resource_name, app_password=password)
+
+    @staticmethod
+    def get_bot_file_encryption_key():
+        """
+
+        :return:
+        """
+
+        # Pulled out of create_app, which is the only place that performs this call
+        # TODO: Move to class for dev.botframework.com calls?
+        response = requests.get('https://dev.botframework.com/api/misc/botFileEncryptionKey')
+
+        # TODO: Tear this out? This method should probably fail gracefully, is a secret needed when the bot is provisioned?
+        # Can't a user create a new secret and then re-encrypt the bot file?
+        if response.status_code not in [200]:
+            raise CLIError('Unable to provision a bot file encryption key. Please try again.')
+        return response.text[1:-1]
+
+    @staticmethod
+    def update(client, parameters, resource_group_name):
+        try:
+            return client.update(
+                resource_group_name=resource_group_name,
+                resource_name=parameters.name,
+                **(parameters.__dict__)
+            )
+        except AttributeError:
+            return None
+
+    @staticmethod
+    def _process_parameters(parameter_lists):
+        def _try_parse_json_object(value):
+            try:
+                parsed = shell_safe_json_parse(value)
+                return parsed.get('parameters', parsed)
+            except CLIError:
+                return None
+
+        parameters = {}
+        for params in parameter_lists or []:
+            for item in params:
+                param_obj = _try_parse_json_object(item)
+                if param_obj:
+                    parameters.update(param_obj)
+
+        return parameters
