+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/custom.py
@@ -6,25 +6,19 @@
 import json
 import os
 import requests
-import re
 import shutil
-import string
 # TODO: Evalulate the use of requests and urllib3, where urllib3 is a dependency of requests
 import urllib3
-import random
 import zipfile
 
-from .auth.converged_app import ConvergedApp
+from .auth import converged_app
+from .bot_template_deployer import BotTemplateDeployer
+from .web_app_operations import WebAppOperations
+from .bot_json_formatter import BotJsonFormatter
 
 from knack.util import CLIError
 from knack.log import get_logger
-from azure.cli.core._profile import Profile
-from azure.cli.command_modules.botservice._webutils import (
-    deploy_arm_template,
-    enable_zip_deploy,
-    get_app_settings,
-    _get_site_credential,
-    _get_scm_url)
+from azure.cli.command_modules.botservice._webutils import enable_zip_deploy
 from azure.mgmt.botservice.models import (
     Bot,
     BotProperties,
@@ -33,55 +27,7 @@ from azure.mgmt.botservice.models import (
     ConnectionSettingParameter,
     Sku)
 
-try:
-    # Try importing Python 3 urllib.parse
-    from urllib.parse import urlsplit
-except ImportError:
-    # If urllib.parse was not imported, use Python 2 module urlparse
-    from urlparse import urlsplit  # pylint: disable=import-error
-
 logger = get_logger(__name__)
-
-
-def get_bot_site_name(endpoint):
-    split_parts = urlsplit(endpoint)
-    return str(split_parts.netloc.split('.', 1)[0])
-
-
-def _get_app_insights_location(key):
-    region_map = {
-        'australiaeast': 'southeastasia',
-        'australiacentral': 'southeastasia',
-        'australiacentral2': 'southeastasia',
-        'australiasoutheast': 'southeastasia',
-        'eastasia': 'southeastasia',
-        'southeastasia': 'westus',
-        'eastus': 'eastus',
-        'eastus2': 'eastus',
-        'southcentralus': 'southcentralus',
-        'westcentralus': 'westus2',
-        'westus': 'westus2',
-        'westus2': 'westus2',
-        'brazilsouth': 'southcentralus',
-        'centralus': 'southcentralus',
-        'northcentralus': 'southcentralus',
-        'japanwest': 'southeastasia',
-        'japaneast': 'southeastasia',
-        'southindia': 'southeastasia',
-        'centralindia': 'southeastasia',
-        'westindia': 'southeastasia',
-        'canadacentral': 'southcentralus',
-        'canadaeast': 'eastus',
-        'koreacentral': 'southeastasia',
-        'koreasouth': 'southeastasia',
-        'northeurope': 'northeurope',
-        'westeurope': 'westeurope',
-        'uksouth': 'westeurope',
-        'ukwest': 'westeurope',
-        'francecentral': 'westeurope',
-        'francesouth': 'westeurope'
-    }
-    return region_map[key]
 
 
 def create(cmd, client, resource_group_name, resource_name, kind, description=None, display_name=None,
@@ -111,7 +57,7 @@ def create(cmd, client, resource_group_name, resource_name, kind, description=No
 
     # If a Microsoft application id was not provided, provision one for the user
     if not msa_app_id:
-        msa_app_id, password = ConvergedApp.provision(resource_name)
+        msa_app_id, password = converged_app.ConvergedApp.provision(resource_name)
         logger.warning('Microsoft application provisioning successful.')
 
     logger.warning('Provisioning bot...')
@@ -144,155 +90,8 @@ def create(cmd, client, resource_group_name, resource_name, kind, description=No
         )
     # Web app and function bots require deploying custom ARM templates, we do that in a separate method
     else:
-        return create_app(cmd, client, resource_group_name, resource_name, description, kind, msa_app_id, password,
+        return BotTemplateDeployer.create_app(cmd, client, resource_group_name, resource_name, description, kind, msa_app_id, password,
                           storageAccountName, location, sku_name, appInsightsLocation, language, version)
-
-
-def create_app(cmd, client, resource_group_name, resource_name, description, kind, appid, password, storageAccountName,
-               location, sku_name, appInsightsLocation, language, version):
-
-    # Based on sdk version, language and kind, select the appropriate zip url containing starter bot source
-    if version == 'v3':
-        if kind == 'function':
-            template_name = 'functionapp.template.json'
-            if language == 'Csharp':
-                zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/csharp-abs-functions_emptybot.zip'
-            else:
-                zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/node.js-abs-functions_emptybot_funcpack.zip'  # pylint: disable=line-too-long
-
-        else:
-            kind = 'sdk'
-            template_name = 'webapp.template.json'
-            if language == 'Csharp':
-                zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/csharp-abs-webapp_simpleechobot_precompiled.zip'  # pylint: disable=line-too-long
-            else:
-                zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/node.js-abs-webapp_hello-chatconnector.zip'  # pylint: disable=line-too-long
-    elif version == 'v4':
-        if kind == 'function':
-            raise CLIError('Function bot creation is not supported for v4 bot sdk.')
-
-        else:
-            kind = 'sdk'
-            template_name = 'webappv4.template.json'
-            if language == 'Csharp':
-                zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/csharp-abs-webapp-v4_echobot_precompiled.zip'  # pylint: disable=line-too-long
-            else:
-                zip_url = 'https://connectorprod.blob.core.windows.net/bot-packages/node.js-abs-webapp-v4_echobot.zip'  # pylint: disable=line-too-long
-
-    # Storage prep
-    create_new_storage = False
-    if not storageAccountName:
-        create_new_storage = True
-        storageAccountName = re.sub(r'[^a-z0-9]', '', resource_name[:10] +
-                                    ''.join(random.choice(string.ascii_lowercase + string.digits) for _ in range(4)))
-        site_name = re.sub(r'[^a-z0-9]', '', resource_name[:15] +
-                           ''.join(random.choice(string.ascii_lowercase + string.digits) for _ in range(4)))
-
-    # Application insights prep
-    appInsightsLocation = _get_app_insights_location(location.lower().replace(' ', ''))
-
-    # ARM Template parameters
-    paramsdict = {
-        "location": location,
-        "kind": kind,
-        "sku": sku_name,
-        "siteName": site_name,
-        "appId": appid,
-        "appSecret": password,
-        "storageAccountResourceId": "",
-        "serverFarmId": "/subscriptions/{0}/resourceGroups/{1}/providers/Microsoft.Web/serverfarms/{2}".format(
-            client.config.subscription_id, resource_group_name, resource_name),
-        "zipUrl": zip_url,
-        "createNewStorage": create_new_storage,
-        "storageAccountName": storageAccountName,
-        "botEnv": "prod",
-        "useAppInsights": True,
-        "appInsightsLocation": appInsightsLocation,
-        "createServerFarm": True,
-        "serverFarmLocation": location.lower().replace(' ', ''),
-        "azureWebJobsBotFrameworkDirectLineSecret": "",
-        "botId": resource_name
-    }
-    if description:
-        paramsdict['description'] = description
-    if template_name == 'webappv4.template.json':
-        bot_encryption_key = get_bot_file_encryption_key()
-        paramsdict['botFileEncryptionKey'] = bot_encryption_key
-    params = {k: {'value': v} for k, v in paramsdict.items()}
-
-    # TODO concate 'templates' to file path
-    dir_path = os.path.dirname(os.path.realpath(__file__))
-
-    # Deploy ARM template
-    deploy_result = deploy_arm_template(
-        cli_ctx=cmd.cli_ctx,
-        resource_group_name=resource_group_name,
-        template_file=os.path.join(dir_path, template_name),
-        parameters=[[json.dumps(params)]],
-        deployment_name=resource_name,
-        mode='Incremental'
-    )
-
-    deploy_result.wait()
-    return create_bot_json(cmd, client, resource_group_name, resource_name, app_password=password)
-
-
-def get_bot_file_encryption_key():
-    """
-
-    :return:
-    """
-    
-    # Pulled out of create_app, which is the only place that performs this call
-    # TODO: Move to class for dev.botframework.com calls?
-    response = requests.get('https://dev.botframework.com/api/misc/botFileEncryptionKey')
-
-    # TODO: Tear this out? This method should probably fail gracefully, is a secret needed when the bot is provisioned?
-    # Can't a user create a new secret and then re-encrypt the bot file?
-    if response.status_code not in [200]:
-        raise CLIError('Unable to provision a bot file encryption key. Please try again.')
-    return response.text[1:-1]
-
-
-def update(client, parameters, resource_group_name):
-    try:
-        return client.update(
-            resource_group_name=resource_group_name,
-            resource_name=parameters.name,
-            **(parameters.__dict__)
-        )
-    except AttributeError:
-        return None
-
-
-def create_bot_json(cmd, client, resource_group_name, resource_name, app_password=None, raw_bot_properties=None):
-    if not raw_bot_properties:
-        raw_bot_properties = client.bots.get(
-            resource_group_name=resource_group_name,
-            resource_name=resource_name
-        )
-    if not app_password:
-        site_name = get_bot_site_name(raw_bot_properties.properties.endpoint)
-        app_settings = get_app_settings(
-            cmd=cmd,
-            resource_group_name=resource_group_name,
-            name=site_name
-        )
-        app_password = [item['value'] for item in app_settings if item['name'] == 'MicrosoftAppPassword'][0]
-
-    profile = Profile(cli_ctx=cmd.cli_ctx)
-    return {
-        'type': 'abs',
-        'id': raw_bot_properties.name,
-        'name': raw_bot_properties.properties.display_name,
-        'appId': raw_bot_properties.properties.msa_app_id,
-        'appPassword': app_password,
-        'endpoint': raw_bot_properties.properties.endpoint,
-        'resourceGroup': str(resource_group_name),
-        'tenantId': profile.get_subscription(subscription=client.config.subscription_id)['tenantId'],
-        'subscriptionId': client.config.subscription_id,
-        'serviceName': resource_name
-    }
 
 
 def get_bot(cmd, client, resource_group_name, resource_name, bot_json=None):
@@ -301,7 +100,7 @@ def get_bot(cmd, client, resource_group_name, resource_name, bot_json=None):
         resource_name=resource_name
     )
     if bot_json:
-        return create_bot_json(cmd, client, resource_group_name, resource_name, raw_bot_properties=raw_bot_properties)
+        return BotJsonFormatter.create_bot_json(cmd, client, resource_group_name, resource_name, raw_bot_properties=raw_bot_properties)
 
     return raw_bot_properties
 
@@ -414,32 +213,6 @@ def _prepare_publish_v4(code_dir, proj_file):
         os.chdir(save_cwd)
 
 
-def _prepare_publish_v4(code_dir, proj_file):
-    save_cwd = os.getcwd()
-    os.chdir(code_dir)
-    try:
-        if not os.path.exists(os.path.join('.', 'package.json')):
-            if proj_file is None:
-                raise CLIError('expected --proj-file parameter for csharp v4 project.')
-            with open('.deployment', 'w') as f:
-                f.write('[config]\n')
-                proj_file = proj_file.lower()
-                proj_file = proj_file if proj_file.endswith('.csproj') else proj_file + '.csproj'
-                f.write('SCM_SCRIPT_GENERATOR_ARGS=--aspNetCore {0}\n'.format(find_proj(proj_file)))
-
-        else:
-            # put iisnode.yml and web.config
-            response = requests.get('https://icscratch.blob.core.windows.net/bot-packages/node_v4_publish.zip')
-            with open('temp.zip', 'wb') as f:
-                f.write(response.content)
-            zip_ref = zipfile.ZipFile('temp.zip')
-            zip_ref.extractall()
-            zip_ref.close()
-            os.remove('temp.zip')
-    finally:
-        os.chdir(save_cwd)
-
-
 def publish_app(cmd, client, resource_group_name, resource_name, code_dir=None, proj_file=None, sdk_version='v3'):
     # get the bot and ensure it's not a registration only bot
     raw_bot_properties = client.bots.get(
@@ -455,6 +228,7 @@ def publish_app(cmd, client, resource_group_name, resource_name, code_dir=None, 
     if not os.path.isdir(code_dir):
         raise CLIError('Please supply a valid directory path containing your source code')
     # ensure that the directory contains appropriate post deploy scripts folder
+    # TODO: Find out if PostDeployScripts is langauge specific
     if 'PostDeployScripts' not in os.listdir(code_dir):
         if sdk_version == 'v4':
             # automatically run prepare-publish in case of v4.
@@ -463,22 +237,26 @@ def publish_app(cmd, client, resource_group_name, resource_name, code_dir=None, 
             raise CLIError('Not a valid azure publish directory. Please run prepare-publish.')
 
     zip_filepath = create_upload_zip(code_dir, include_node_modules=False)
-    site_name = get_bot_site_name(raw_bot_properties.properties.endpoint)
+    site_name = WebAppOperations.get_bot_site_name(raw_bot_properties.properties.endpoint)
     # first try to put the zip in clirepo
-    user_name, password = _get_site_credential(cmd.cli_ctx, resource_group_name, site_name, None)
-    scm_url = _get_scm_url(cmd, resource_group_name, site_name, None)
+    user_name, password = WebAppOperations.get_site_credential(cmd.cli_ctx, resource_group_name, site_name, None)
+    scm_url = WebAppOperations.get_scm_url(cmd, resource_group_name, site_name, None)
 
     authorization = urllib3.util.make_headers(basic_auth='{0}:{1}'.format(user_name, password))
     headers = authorization
 
+    # the `clirepo/` folder contains the zipped up source code
     payload = {
         'command': 'rm -rf clirepo',
         'dir': r'site'
     }
     headers['content-type'] = 'application/json'
     response = requests.post(scm_url + '/api/command', data=json.dumps(payload), headers=headers)
+    # TODO: Verify the necessity of this line
     response = requests.put(scm_url + '/api/vfs/site/clirepo/', headers=headers)
+    # TODO: Think about using Response.ok
     check_response_status(response, 201)
+
     headers['content-type'] = 'application/octet-stream'
     with open(zip_filepath, 'rb') as fs:
         zip_content = fs.read()
@@ -486,6 +264,7 @@ def publish_app(cmd, client, resource_group_name, resource_name, code_dir=None, 
 
     output = enable_zip_deploy(cmd, resource_group_name, site_name, 'upload.zip')
     os.remove('upload.zip')
+    # TODO: Examine improving logic for langauge of the bot.
     if os.path.exists(os.path.join('.', 'package.json')):
         payload = {
             'command': 'npm install',
@@ -513,10 +292,10 @@ def download_app(cmd, client, resource_group_name, resource_name, file_save_path
         raise CLIError('The path {0} already exists. Please delete this folder or specify an alternate path'.format(folder_path))  # pylint: disable=line-too-long
     os.mkdir(folder_path)
 
-    site_name = get_bot_site_name(raw_bot_properties.properties.endpoint)
+    site_name = WebAppOperations.get_bot_site_name(raw_bot_properties.properties.endpoint)
 
-    user_name, password = _get_site_credential(cmd.cli_ctx, resource_group_name, site_name, None)
-    scm_url = _get_scm_url(cmd, resource_group_name, site_name, None)
+    user_name, password = WebAppOperations.get_site_credential(cmd.cli_ctx, resource_group_name, site_name, None)
+    scm_url = WebAppOperations.get_scm_url(cmd, resource_group_name, site_name, None)
 
     authorization = urllib3.util.make_headers(basic_auth='{0}:{1}'.format(user_name, password))
     headers = authorization
@@ -542,14 +321,17 @@ def download_app(cmd, client, resource_group_name, resource_name, file_save_path
     zip_ref.extractall(folder_path)
     zip_ref.close()
     os.remove(download_path)
+    # TODO: Examine cases where PostDeployScripts, deploy.cmd, etc. do not exist.
     if (os.path.exists(os.path.join(folder_path, 'PostDeployScripts', 'deploy.cmd.template')) and
             os.path.exists(os.path.join(folder_path, 'deploy.cmd'))):
         shutil.copyfile(os.path.join(folder_path, 'deploy.cmd'),
                         os.path.join(folder_path, 'PostDeployScripts', 'deploy.cmd.template'))
     # if the bot contains a bot
+    # TODO: If there is only one bot file, that is the bot file.
+    # TODO: If there are more than one bot file, the user must disambiguate before continuing
     bot_file_path = os.path.join(folder_path, '{0}.bot'.format(resource_name))
     if os.path.exists(bot_file_path):
-        app_settings = get_app_settings(
+        app_settings = WebAppOperations.get_app_settings(
             cmd=cmd,
             resource_group_name=resource_group_name,
             name=site_name
